% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_miles.R
\name{bike_miles}
\alias{bike_miles}
\title{Caculate total miles biked}
\usage{
bike_miles(distance, count)
}
\arguments{
\item{distance}{Number of miles biked}

\item{count}{Number of times biked}
}
\description{
Caculate total miles biked
}
